requirements.txt
mlx
tqdm
python-mnist
numpy
mlflow
torch
torchvision